pyproject.toml
[tool.poetry]
name = "discord-bot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
grpcio = "^1.60.0"
discord-py = "^2.3.2"


[tool.poetry.group.dev.dependencies]
grpcio-tools = "^1.60.0"
protoletariat = "^3.2.19"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
